Import CenterEdgeOps legacy assets (DATA-626) (#15)
Motivation
---
It has been a longstanding issue that legacy Operations resources in AWS were not imported into this terraform repo. This is the beginning of efforts to import those resources and manage them in terraform going forward.

Modifications
---
Added to terraform:

- The load balancer and related items that serve traffic for Grafana and Jarvis.
- The Postgres RDS instance that hosts Ops/Data information that is served in QuickSight, Grafana, and Jarvis.
- The permissions, launch templates, VPC, and related items that support Training Virtuals for Ops (the instances themselves are never expected to be included in Terraform)
- Added S3 buckets that are used to store Ops emails and the setup files for the Install team
- The SES and SNS resources that manage incoming Ops emails

NOTE 1: All of the resources added to terraform were imported. No new infrastructure was created as a part of this effort.

NOTE 2: Remaining resources to import include:

- A lambda function that processes incoming Ops emails.
- A dynamo DB database that hosts metadata about the processes Ops emails.

It appears as though these resources stopped working in 2021, so some effort will need to go in to updating them to a functioning state.

https://centeredge.atlassian.net/browse/DATA-626