Import ops lambda function and resources to terraform (DATA-632) (#17)
Motivation
---

This is part two to [DATA-626](https://centeredge.atlassian.net/browse/DATA-626) where we are completing the migration by importing the emailHandler lambda function and its associated resources (including the DynamoDB database that stores the data).

Modifications
---
* Added the `emailHandler` lambda function and execution role.
* Added the SNS subscription that serves as the lambda's trigger.
* Added the DynamoDB table where the lambda function stores data.
* Updated the ingress rule for the Office VPN.

Please also note: There is deliberately no source_code_hash or other argument to link the lambda function's code to the infrastructure. I didn't want any potential change to trigger a lambda deployment so as to keep things simple for everyone. Therefore, updates to the lambda function will need to be pushed to production manually (see the function ReadMe for details).

https://centeredge.atlassian.net/browse/DATA-632

[DATA-626]: https://centeredge.atlassian.net/browse/DATA-626?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ